Changed Baud Slower
Added debug signals.

Changed baud rate to 9600 so UART will work.

Reason 115200 failed is because at a clock rate of 10000000, the Ian's
division ends up being 86, when it is 86.8.  That is a 1.7% error,
causing the clocks to be off by 17% at the end of the byte.  Long term
fix is to run the whole system off 100 MHz clock.